usb_acm: configured usb regulator gpio pin to output in soc_init
When device is booted up with usb unplugged, bootloader does not
perform enable usb functionality which configures the usb regulator
gpio pin to output. This was leading to acm port not enumerating
when usb is plugged as, we are writing to the usb regulator pin
on every usb plug/unplug event.